Split Stratify into its own file.
